% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BinarySegmentation.R
\name{BINSEG}
\alias{BINSEG}
\title{Title}
\usage{
BINSEG(data, segments)
}
\arguments{
\item{data}{datae data}

\item{segments}{the number of segments}
}
\value{
{estimated signal, estimated means, estimated changepoints}
}
\description{
Title
}
