% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BinarySegmentation.R
\name{bs.helper}
\alias{bs.helper}
\title{Title}
\usage{
bs.helper(data, former.segments = NULL, segments)
}
\arguments{
\item{data}{observations}

\item{former.segments}{intervals to be tested}

\item{segments}{critical value used in testing}
}
\value{
finsal.segments (intervals to test in next iteration)
}
\description{
Title
}
