% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BinarySegmentation.R
\name{bs.model}
\alias{bs.model}
\title{Title}
\usage{
bs.model(data, interval)
}
\arguments{
\item{data}{observations}

\item{interval}{intervals to partition observations}
}
\value{
a list of est.signal and est.means
}
\description{
Title
}
