% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BinarySegmentation.R
\name{bs_stat}
\alias{bs_stat}
\title{Title}
\usage{
bs_stat(data, lower, upper)
}
\arguments{
\item{data}{observations}

\item{lower}{lower open interval bound}

\item{upper}{upper open interval bound}
}
\value{
vector containing all CUSUM values for all points in the interval
}
\description{
Title
}
